requirement.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.24.3
tldextract==3.4.4
python-whois==0.8.0
shap==0.42.1